+++ src/components/InfoCard/InfoCard.vue
@@ -1,8 +1,9 @@
 <template>
   <div :class="classList">
+    <div :class="brandBarClassList" v-if="brandBar" />
     <div class="ph-bg-white">
-      <div v-if="heading" class="ph-bg-grey5_o15 ph-w-full ph-relative">
-        <div class="ph-p-5">
+      <div v-if="heading" class="ph-bg-grey10 ph-w-full ph-relative">
+        <div class="ph-p-5" :class="!brandBar ? 'ph-pl-9' : ''">
           <PBrandGradientText>
             <span class="ph-text-2xl ph-font-bold">
               {{ heading }}
@@ -16,7 +17,13 @@
         </div>
       </div>
 
-      <div class="ph-w-full ph-p-5" :class="`ph-bg-${contentBg}`">
+      <div
+        class="ph-w-full ph-p-5 ph-pr-10"
+        :class="[
+          `ph-bg-${contentBg}`,
+          !brandBar ? 'ph-pl-9' : '',
+        ]"
+      >
         <slot />
       </div>
     </div>  
@@ -60,18 +67,34 @@ export default Vue.extend({
       type: Boolean as PropType<boolean>,
       default: true,
     },
+    
+    brandBar: {
+      type: Boolean as PropType<boolean>,
+      default: false,
+    },
   },
 
   computed: {
     classList():string[] {
       return [
+        'ph-relative',
+        'ph-rounded-lg',
+        this.brandBar ? 'ph-pl-4' : '',
+        'ph-overflow-hidden',
+        this.shadow ? 'ph-shadow' : '',
+      ]
+    },
+    
+    brandBarClassList():string[] {
+      return [
         'ph-bg-gradient-to-t',
         'ph-from-brand2',
         'ph-to-brand1',
-        'ph-rounded-lg',
-        'ph-pl-4',
-        'ph-overflow-hidden',
-        this.shadow ? 'ph-shadow' : '',
+        'ph-w-4',
+        'ph-h-full',
+        'ph-top-0',
+        'ph-left-0',
+        'ph-absolute',
       ]
     },
     
@@ -81,7 +104,7 @@ export default Vue.extend({
         'ph-absolute',
         'ph-w-16',
         'ph-h-16',
-        'ph-right-10',
+        'ph-right-5',
         'ph-top-1/2',
         'ph-mt-1',
         'ph-rounded-full',