+++ src/components/index.ts
@@ -3,8 +3,10 @@ import PIcon from './Icon';
 import PButton from './Button';
 import PSkipLink from './SkipLink';
 import PAccordion from './Accordion';
+import PForm from './Form';
 import PInput from './Input';
 import PLabel from './Label';
+import PInputPhone from './InputPhone';
 import PInputText from './InputText';
 import PAutoComplete from './Autocomplete';
 import PDialog from './Dialog';
@@ -13,6 +15,8 @@ import PCard from './Card';
 import PRate from './Rate';
 import PDateInput from './DateInput';
 import PCheckbox from './Checkbox';
+import PInputRadio from './InputRadio';
+import PRadioGroup from './RadioGroup';
 import PMessage from './Message';
 import PToast from './Toast';
 import PList from './List';
@@ -38,6 +42,7 @@ const components: PhotonComponents = {
   PAccordion,
   PAutoComplete,
   PIcon,
+  PInputPhone,
   PInputText,
   PInput,
   PLabel,
@@ -58,6 +63,9 @@ const components: PhotonComponents = {
   PLi,
   PInfoCard,
   PBrandGradientText,
+  PForm,
+  PInputRadio,
+  PRadioGroup
 };
 
 export type PhotonComponentKeys = keyof PhotonComponents